Implemented Excel Support & Random Colors
Random colors are within the script in the charts html file. I think this could be done in an easier way, but this was the first solution I stumbled upon. The views file within the charts app is doing the heavy lifting for reading excel files and creating many variables for the charts. I'm not sure if this should all be together in one app